Route Zano address handling through ChainRegistry
Adds addressHandlers, domainAddressHandlers and addressValidator hooks.
AddressHandlerZano, AddressHandlerZanoAlias, ZanoAliasResolver and
ZanoAddressValidator move to the chain.zano package; the four dispatch
sites (plain handlers, factory domain handlers, EnterAddressViewModel,
AddressValidatorFactory) resolve them via the registry.
