+++ walletkit/src/main/java/io/horizontalsystems/walletkit/chain/zano/ZanoAliasResolver.kt
@@ -1,4 +1,4 @@
-package io.horizontalsystems.walletkit.core.address
+package io.horizontalsystems.walletkit.chain.zano
 
 import com.google.gson.JsonObject
 import com.google.gson.JsonParser
